Bug #1189611 drm/sprd: sharkl5pro cabc support 10bit backlight
[root cause  ] cabc backlight 10bit support
[changes     ] cabc
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:sharkl5pro
[test case]:display normally
[reviewers   ] leon.he
[change_type ] feature_bugfix
[tag_product ] common

Sharkl5pro cabc support 10bit backlight.

+++ drivers/gpu/drm/sprd/cabc/cabc_definition.h
@@ -24,14 +24,14 @@
 #define PRE_PRE 1
 
 struct bl_out_tag {
-	u8 cur;
-	u8 pre;
-	u8 pre2;
-	u8 pre3;
+	u16 cur;
+	u16 pre;
+	u16 pre2;
+	u16 pre3;
 	int cur_fix_ui;
-	u8 pre_fix_ui;
+	u16 pre_fix_ui;
 	int cur_fix_video;
-	u8 pre_fix_video;
+	u16 pre_fix_video;
 };
 
 struct cabc_context_tag {
